Fix init_test/websocket build and init compiler warnings
Add websocket/find_path Boost includes; wire websocket and init tests in root CMake.
init tests call ProcessCommand via test-only private exposure; tighten init CMake link set.
Silence genesis dead locals and network_init assert signedness; minor test CMake and coverage edits.

+++ src/init/network_init.cc
@@ -1836,7 +1836,7 @@ void NetworkInit::SendJoinElectTransaction() {
             CreateContribution(req);
         }
 #ifndef NDEBUG
-        assert(req->verify_vec_size() >= t);
+        assert(static_cast<uint32_t>(req->verify_vec_size()) >= t);
 #endif
         SETH_DEBUG("First time sending join_elect or no valid g2_req in consensus block, "
             "including g2_req in transaction. verify_vec_size: %d", req->verify_vec_size());
